NodeDefExpressionValidator: splitted nodeDefExpressionValidator into several files
