+++ core/survey/nodeDefExpressionValidator/index.js
@@ -1,0 +1,1 @@
+export { validate } from './nodeDefExpressionValidator'
